Create a common class for api exceptions handling (ExceptionApiHandler); move handlers from controller to ExceptionApiHandler; add handler for HttpMessageNotReadableException
