--- calculator/src/main/java/org/smileksey/calculator/utils/validation/GenderSubset.java
@@ -1,24 +1,0 @@
-package org.smileksey.calculator.utils.validation;
-
-import org.smileksey.calculator.dto.enums.Gender;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-//FIXME
-@Target({METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER, TYPE_USE})
-@Retention(RUNTIME)
-@Documented
-@Constraint(validatedBy = GenderSubSetValidator.class)
-public @interface GenderSubset {
-    Gender[] anyOf();
-    String message() default "must be any of {anyOf}";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
